fix some pir ut in test/amp
