+++ test/amp/test_amp_list.py
@@ -16,6 +16,7 @@ import unittest
 
 import paddle
 from paddle.base import core
+from paddle.pir_utils import test_with_pir_api
 from paddle.static.amp import AutoMixedPrecisionLists, fp16_lists
 
 
@@ -113,6 +114,7 @@ class TestAMPList(unittest.TestCase):
                 self.assertEqual(out2.dtype, core.DataType.FLOAT32)
                 self.assertEqual(out3.dtype, core.DataType.FLOAT32)
 
+    @test_with_pir_api
     def test_apis(self):
         def _run_check_dtype():
             fp16_lists.check_amp_dtype(dtype="int64")
@@ -141,14 +143,30 @@ class TestAMPList(unittest.TestCase):
             self.assertEqual(
                 fp16_lists.get_low_precision_dtypestr(dtype), dtype
             )
-        self.assertEqual(
-            fp16_lists.get_low_precision_dtypestr(core.VarDesc.VarType.FP16),
-            "float16",
-        )
-        self.assertEqual(
-            fp16_lists.get_low_precision_dtypestr(core.VarDesc.VarType.BF16),
-            "bfloat16",
-        )
+        if paddle.framework.use_pir_api():
+            self.assertEqual(
+                fp16_lists.get_low_precision_dtypestr(core.DataType.FLOAT16),
+                "float16",
+            )
+        else:
+            self.assertEqual(
+                fp16_lists.get_low_precision_dtypestr(
+                    core.VarDesc.VarType.FP16
+                ),
+                "float16",
+            )
+        if paddle.framework.use_pir_api():
+            self.assertEqual(
+                fp16_lists.get_low_precision_dtypestr(core.DataType.BFLOAT16),
+                "bfloat16",
+            )
+        else:
+            self.assertEqual(
+                fp16_lists.get_low_precision_dtypestr(
+                    core.VarDesc.VarType.BF16
+                ),
+                "bfloat16",
+            )
 
         def _run_get_dtypestr():
             fp16_lists.get_low_precision_dtypestr(dtype="int64")
