+++ test/amp/test_amp_o2_embedding_model_deprecated.py
@@ -119,49 +119,50 @@ class TestUnittedEmbedding(AmpTestBase):
         x = np.random.randint(1, 64, size=[1, 32]).astype("int64")
         return x
 
-    def test_pir_compare_o1_and_o2_master_grad(self):
-        def _run(data, level, use_promote=False):
-            with paddle.pir_utils.IrGuard():
-                startup = paddle.static.Program()
-                main = paddle.static.Program()
-                with paddle.static.program_guard(main, startup):
-                    model, optimizer, scaler = build_unitted_embedding_model(
-                        use_amp=True,
-                        amp_dtype="float16",
-                        amp_level=level,
-                        use_promote=use_promote,
-                    )
-                    model.train()
-                    with paddle.amp.auto_cast(
-                        enable=True,
-                        dtype='float16',
-                        level=level,
-                        use_promote=use_promote,
-                    ):
-                        x = paddle.static.data('x', [None, 32], 'int64')
-                        out = model(x)
-                        loss = paddle.mean(out)
-                    scaled = scaler.scale(loss)
-                    scaler.minimize(optimizer, scaled)
+    def test_compare_o1_and_o2_master_grad(self):
+        def _run(place, exe, x_np, max_iters, level):
+            (
+                main_program,
+                startup_program,
+                optimizer,
+                feed_vars,
+                fetch_vars,
+            ) = build_unitted_embedding_model(
+                True,
+                "float16",
+                level,
+            )
 
-                    if paddle.is_compiled_with_cuda():
-                        place = paddle.CUDAPlace(0)
-                    elif paddle.device.is_compiled_with_xpu():
-                        place = paddle.device.XPUPlace(0)
-                    else:
-                        raise ValueError("Only support CUDA or XPU Place.")
-                    exe = paddle.static.Executor(place)
-                    exe.run(startup)
-                    exe.run(
-                        main,
-                        feed={
-                            'x': data,
-                        },
-                        fetch_list=[loss],
-                    )
+            seed = 0
+            paddle.seed(seed)
+            np.random.seed(seed)
+            random.seed(seed)
 
-        x_data = self._generate_feed_x()
-        _run(x_data, "O2")
+            losses = self.run_program(
+                main_program,
+                startup_program,
+                optimizer,
+                feed_vars,
+                fetch_vars,
+                place,
+                exe,
+                x_np,
+                max_iters,
+                "float16",
+                level,
+            )
+            return losses
+
+        max_iters = 5
+        x = self._generate_feed_x()
+        if paddle.is_compiled_with_cuda():
+            place = paddle.CUDAPlace(0)
+        elif paddle.device.is_compiled_with_xpu():
+            place = paddle.device.XPUPlace(0)
+        else:
+            raise ValueError("Only support CUDA or XPU Place.")
+        exe = paddle.static.Executor(place)
+        losses_o2 = _run(place, exe, x, max_iters, 'O2')
 
 
 if __name__ == "__main__":
